requirements.txt
numpy~=1.22.0
matplotlib~=3.5.0
opencv-python~=4.5.5.62
pathfinding~=1.0.1
pillow~=8.4.0